+++ src/legacy/ui/public/agg_types/param_types/field.ts
@@ -25,7 +25,7 @@ import { FieldParamEditor } from '../../vis/editors/default/controls/field';
 import { BaseParamType } from './base';
 import { toastNotifications } from '../../notify';
 import { propFilter } from '../filter';
-import { Field, FieldListInterface } from '../../index_patterns';
+import { Field, IFieldList } from '../../index_patterns';
 
 const filterByType = propFilter('type');
 
@@ -111,7 +111,7 @@ export class FieldParamType extends BaseParamType {
   /**
    * filter the fields to the available ones
    */
-  getAvailableFields = (fields: FieldListInterface) => {
+  getAvailableFields = (fields: IFieldList) => {
     const filteredFields = fields.filter((field: Field) => {
       const { onlyAggregatable, scriptable, filterFieldTypes } = this;
