+++ src/plugins/data/public/index_patterns/index.ts
@@ -43,7 +43,7 @@ export const indexPatterns = {
 };
 
 export { IndexPatternsService } from './index_patterns_service';
-export { Field, FieldList, FieldListInterface } from './fields';
+export { Field, FieldList, IFieldList } from './fields';
 
 // TODO: figure out how to replace IndexPatterns in get_inner_angular.
 export { IndexPattern, IndexPatterns, IndexPatternsContract } from './index_patterns';
