+++ src/plugins/data/public/index_patterns/index_patterns/index_patterns.ts
@@ -28,6 +28,7 @@ import {
 import { createIndexPatternCache } from './_pattern_cache';
 import { IndexPattern } from './index_pattern';
 import { IndexPatternsApiClient, GetFieldsOptions } from './index_patterns_api_client';
+import { FieldList } from '..';
 
 const indexPatternCache = createIndexPatternCache();
 
@@ -147,4 +148,6 @@ export class IndexPatterns {
   };
 }
 
-export type IndexPatternsContract = PublicMethodsOf<IndexPatterns>;
+export type IndexPatternsContract = PublicMethodsOf<IndexPatterns> & {
+  FieldList: typeof FieldList;
+};
