+++ src/test_utils/public/stub_index_pattern.js
@@ -21,15 +21,13 @@ import sinon from 'sinon';
 // TODO: We should not be importing from the data plugin directly here; this is only necessary
 // because it is one of the few places that we need to access the IndexPattern class itself, rather
 // than just the type. Doing this as a temporary measure; it will be left behind when migrating to NP.
-import { IndexPattern } from '../../legacy/core_plugins/data/public/';
 import {
   FieldList,
-  getRoutes,
-  formatHitProvider,
-  flattenHitWrapper,
 } from 'ui/index_patterns';
 import {
   FIELD_FORMAT_IDS,
+  IndexPattern,
+  indexPatterns,
 } from '../../plugins/data/public';
 
 import { getFieldFormatsRegistry } from './stub_field_formats';
@@ -49,11 +47,11 @@ export default  function StubIndexPattern(pattern, getConfig, timeField, fields,
   this.getSourceFiltering = sinon.stub();
   this.metaFields = ['_id', '_type', '_source'];
   this.fieldFormatMap = {};
-  this.routes = getRoutes();
+  this.routes = indexPatterns.getRoutes();
 
   this.getComputedFields = IndexPattern.prototype.getComputedFields.bind(this);
-  this.flattenHit = flattenHitWrapper(this, this.metaFields);
-  this.formatHit = formatHitProvider(this, registeredFieldFormats.getDefaultInstance(FIELD_FORMAT_IDS.STRING));
+  this.flattenHit = indexPatterns.flattenHitWrapper(this, this.metaFields);
+  this.formatHit = indexPatterns.formatHitProvider(this, registeredFieldFormats.getDefaultInstance(FIELD_FORMAT_IDS.STRING));
   this.fieldsFetcher = { apiClient: { baseUrl: '' } };
   this.formatField = this.formatHit.formatField;
